# openapi.yaml
openapi: 3.1.0
info:
  title: SecureShield DRM API
  version: "1.0.0"
  description: |
    Digital Rights Management backend for SecureShield.

    ## Authentication
    Admin endpoints require the `X-Admin-Key` header.
    Public endpoints (verify-license, analytics, offline-token verification) require no header.

    ## Base URL
    Development: `http://localhost:8001`
servers:
  - url: http://localhost:8001
    description: Local development
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Static admin API key. Set via `ADMIN_API_KEY` environment variable.
    SuperAdminKey:
      type: apiKey
      in: header
      name: X-Super-Admin-Key
      description: Super-admin key for tenant management. Set via `SUPER_ADMIN_KEY` environment variable.
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant slug. Combined with X-Admin-Key for tenant-scoped endpoints.
  schemas:
...
    IsoTimestamp:
      type: string
      format: date-time
      example: "2026-02-20T14:30:00Z"
...
    Error:
      type: object
      properties:
        detail: { type: string }
...
    Tenant:
      type: object
      properties:
        id:           { type: integer }
        name:         { type: string, example: "Acme Corp" }
        slug:         { type: string, example: "acme-corp" }
        plan:         { type: string, enum: [starter, pro, enterprise], example: "starter" }
        max_licenses: { type: integer, example: 10 }
        max_vault_mb: { type: integer, example: 100 }
        is_active:    { type: boolean, example: true }
        created_at:   { $ref: "#/components/schemas/IsoTimestamp" }
...
paths:
...
  /superadmin/tenants:
    post:
      tags: [superadmin]
      summary: Create a new tenant
      operationId: createTenant
      security:
        - SuperAdminKey: []
      parameters:
        - { name: name,         in: query, required: true,  schema: { type: string } }
        - { name: slug,         in: query, required: true,  schema: { type: string, example: "acme-corp" } }
        - { name: admin_key,    in: query, required: true,  schema: { type: string, description: "Plain-text key — bcrypt-hashed before storage" } }
        - { name: plan,         in: query, required: false, schema: { type: string, enum: [starter, pro, enterprise], default: "starter" } }
        - { name: max_licenses, in: query, required: false, schema: { type: integer, default: 10 } }
        - { name: max_vault_mb, in: query, required: false, schema: { type: integer, default: 100 } }
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409": { description: Slug already taken, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

    get:
      tags: [superadmin]
      summary: List all tenants
      operationId: listTenants
      security:
        - SuperAdminKey: []
      responses:
        "200":
          description: All tenants ordered by creation date
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tenant" }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /superadmin/tenants/{slug}:
    patch:
      tags: [superadmin]
      summary: Update a tenant (plan, limits, active status)
      operationId: updateTenant
      security:
        - SuperAdminKey: []
      parameters:
        - { name: slug,         in: path,  required: true,  schema: { type: string } }
        - { name: plan,         in: query, required: false, schema: { type: string, enum: [starter, pro, enterprise] } }
        - { name: max_licenses, in: query, required: false, schema: { type: integer } }
        - { name: max_vault_mb, in: query, required: false, schema: { type: integer } }
        - { name: is_active,    in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:    { type: string }
                  updated: { type: boolean }
        "401": { description: Unauthorized }
        "404": { description: Tenant not found }

    delete:
      tags: [superadmin]
      summary: Delete a tenant permanently
      operationId: deleteTenant
      security:
        - SuperAdminKey: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: deleted }
                  slug:   { type: string }
        "401": { description: Unauthorized }
        "404": { description: Tenant not found }

  # ── Tenant-Scoped Admin Endpoints ─────────────────────────────────────────
...